% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_call2.R
\name{euler_call2}
\alias{euler_call2}
\title{European Call Function}
\usage{
euler_call2(X0, mu, sigma, Dt = 0.1, t = 0, T = 1, N = 1000, K = 0,
  plt = FALSE)
}
\arguments{
\item{X0}{Initial value.}

\item{mu}{Drift coeficient.}

\item{sigma}{Diffusion coefficient.}

\item{Dt}{Step size.}

\item{t}{Initial time period.}

\item{T}{Final time period.}

\item{N}{Number of simulations.}

\item{K}{Strike price.}

\item{plt}{Plot? Defaults to FALSE.}
}
\description{
#' This function allows you to simulate a call price using the Euler method.
}
\examples{
euler_call2(307.65, 0.75, 0.3, 0.001, 0, 1, 10000)

}
\author{
Fernando Teixeira
}
\keyword{Euler}
\keyword{method}
