% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston_ea_bk.R
\name{hestonea}
\alias{hestonea}
\title{FUNCTION_TITLE}
\usage{
hestonea(S, X, r, v, theta, rho, k, sigma, t = 0, tau = 1)
}
\arguments{
\item{S}{PARAM_DESCRIPTION}

\item{X}{PARAM_DESCRIPTION}

\item{r}{PARAM_DESCRIPTION}

\item{v}{PARAM_DESCRIPTION}

\item{theta}{PARAM_DESCRIPTION}

\item{rho}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION}

\item{sigma}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION, Default: 0}

\item{tau}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{rchisq}},\code{\link[stats]{rnorm}}
}
