% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston_ea_bk.R
\name{intv}
\alias{intv}
\title{FUNCTION_TITLE}
\usage{
intv(n, cf, v_t)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{cf}{PARAM_DESCRIPTION}

\item{v_t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{integrate}},\code{\link[stats]{uniroot}},\code{\link[stats]{runif}}
}
