% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milstein_call.R
\name{milstein_call}
\alias{milstein_call}
\title{European Call Function}
\usage{
milstein_call(X0, mu, sigma, Dt, t, T, N, K = 0, plt = F)
}
\arguments{
\item{X0}{Initial value.}

\item{mu}{Drift coeficient.}

\item{sigma}{Diffusion coefficient.}

\item{Dt}{Step size.}

\item{t}{Initial time period.}

\item{T}{Final time period.}

\item{N}{Number of simulations.}

\item{K}{Strike price.}

\item{plt}{Plot? Defaults to FALSE.}
}
\description{
This function allows you to simulate a call price using the Milstein method.
}
\examples{
milstein_call(307.65, 0.75, 0.3, 0.001, 0, 1, 10000, 300)


}
\author{
Fernando Teixeira
}
\keyword{method}
\keyword{milstein}
