% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heston_ea_bk.R
\name{phi_heston}
\alias{phi_heston}
\title{FUNCTION_TITLE}
\usage{
phi_heston(a, v0, v_t, d)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{v0}{PARAM_DESCRIPTION}

\item{v_t}{PARAM_DESCRIPTION}

\item{d}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[Bessel]{BesselI}}
}
